frontend/requirements.txt
web3==6.20.2
python-dotenv==1.0.1
Flask==3.0.3
Werkzeug==3.0.3
flask_cors==4.0.1